Pull makeRequest into separate file (#1915)
